+++ corehq/apps/api/resources/v0_3.py
@@ -88,40 +88,3 @@ class CommCareCaseResource(HqBaseResource, DomainSpecificResourceMixin):
         resource_name = 'case'
         list_allowed_methods = ['get']
         detail_allowed_methods = ['get']
-
-
-class XFormInstanceResource(HqBaseResource, DomainSpecificResourceMixin):
-    id = fields.CharField(attribute='form_id', readonly=True, unique=True)
-
-    form = fields.DictField(attribute='form_data')
-    type = fields.CharField(attribute='type')
-    version = fields.CharField(attribute='version')
-    uiversion = fields.CharField(attribute='uiversion')
-    metadata = fields.DictField(attribute='metadata', null=True)
-    received_on = fields.DateTimeField(attribute="received_on")
-    md5 = fields.CharField(attribute='xml_md5')
-    archived = fields.CharField(readonly=True)
-
-    def dehydrate_archived(self, bundle):
-        return bundle.obj.is_archived
-
-    def detail_uri_kwargs(self, bundle_or_obj):
-        return {
-            'pk': get_obj(bundle_or_obj).form_id
-        }
-
-    def obj_get(self, bundle, **kwargs):
-        instance_id = kwargs['pk']
-        try:
-            return FormAccessors(kwargs['domain']).get_form(instance_id)
-        except XFormNotFound:
-            raise object_does_not_exist("XFormInstance", instance_id)
-
-    class Meta(CustomResourceMeta):
-        authentication = RequirePermissionAuthentication(Permissions.edit_data)
-        object_class = XFormInstance
-        list_allowed_methods = []
-        detail_allowed_methods = ['get']
-        resource_name = 'form'
-        ordering = ['received_on']
-        serializer = XFormInstanceSerializer(formats=['json'])
